requirements.txt
altair==6.0.0
apimatic-core==0.2.24
apimatic-core-interfaces==0.1.8
apimatic-requests-client-adapter==0.1.10
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
CacheControl==0.14.4
cachetools==7.0.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpickle==4.0.5
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
msgpack==1.1.2
narwhals==2.20.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
paypal-server-sdk==2.2.0
pillow==12.2.0
protobuf==7.34.1
psycopg2-binary==2.9.12
pyarrow==24.0.0
pydeck==0.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.49
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
